requirements.txt
pygame-ce>=2.5.0
torch>=2.0.0
numpy>=1.24.0

# ML Training & Experiment Tracking
tensorboard>=2.14.0

# Model Serving API
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# AWS (S3 model storage + DynamoDB model registry)
boto3>=1.28.0
